Azure/LAB10-Functions-HTTPTrigger/function-code/requirements.txt
azure-functions
azure-functions-worker 